+++ packages/map-template/src/components/Search/Categories/Categories.jsx
@@ -133,14 +133,22 @@ function Categories({ onSetSize, getFilteredLocations, searchFieldRef }) {
             updateScrollButtonsState();
         };
 
+        let resizeObserver;
+
         if (categoriesListRef.current) {
             // Because of timing issue, we need to listen to changes inside div element for categories.
             // Based on that we can handle disabling/enabling scroll buttons.
             // Read more: https://developer.mozilla.org/en-US/docs/Web/API/ResizeObserver
-            const resizeObserver = new ResizeObserver(handleResize)
-            resizeObserver.observe(categoriesListRef.current)
+            resizeObserver = new ResizeObserver(handleResize)
+            resizeObserver.observe(categoriesListRef.current);
             categoriesListRef.current.addEventListener('scroll', updateScrollButtonsState);
         }
+
+        return () => {
+            setActiveCategory();
+            resizeObserver?.disconnect();
+            categoriesListRef.current?.removeEventListener('scroll', updateScrollButtonsState);
+        };
     }, []);
 
     /*